backend/auth-service/src/dto/login/user_login_request_dto.py
from pydantic import BaseModel, Field

class UserLoginRequestDTO(BaseModel):
    username: str = Field(..., max_length=15)
    password: str= Field(..., min_length=3, max_length=32)
    email: str= Field(..., max_length=32)
    
    