backend/auth-service/requirements/prod.txt
fastapi==0.92.0
cassandra-driver==3.25.0
pyjwt==2.6.0
uvicorn==0.20.0
slowapi==0.1.7